Refactor: move git subprocess into helper function
